src/modules/restricao/controller/restricao.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  HttpCode,
  Param,
  Post,
  Put,
  Query,
  Request,
  UseGuards,
} from '@nestjs/common';
import {
  ApiBadRequestResponse,
  ApiCreatedResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiTags,
} from '@nestjs/swagger';
import { activityLog } from '../../../common/activiti-log';
import ACLPolice from '../../../common/decorators/acl.decorator';
import ACLGuard from '../../../common/guards/acl.guard';
import JwtAuthGuard from '../../../common/guards/jwt.guard';
import { ErrorResponse } from '../../../common/responses/error';
import { CreateRestricaoDto } from '../dtos/create.dto';
import { UpdateRestricaoDto } from '../dtos/update.dto';
import Restricao from '../entity/restricao.entity';
import { RestricaoService } from '../service/restricao.service';

@ApiTags('Restricao')
@Controller('restricaos')
export class RestricaoController {
  constructor(private service: RestricaoService) {}

  @Get()
  @HttpCode(200)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Search all Restricao' })
  @ApiOkResponse({
    type: [CreateRestricaoDto],
    description: 'The found Restricao',
  })
  async findAll(): Promise<Restricao[]> {
    return await this.service.findAll();
  }

  @Get('/deleted')
  @HttpCode(200)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Search all deleted Restricao' })
  @ApiOkResponse({
    type: [CreateRestricaoDto],
    description: 'The found deleted Restricao',
  })
  async listDeleted(): Promise<Restricao[]> {
    return await this.service.listDeleted();
  }

  @Get('/andamento')
  @HttpCode(200)
  @ApiOperation({ summary: 'Search all Restricao' })
  // @ApiOkResponse({ type: [any], description: 'The found Restricao' })
  async andamento(): Promise<any[]> {
    return await this.service.findAndamento();
  }

  @Get('/resultado')
  @HttpCode(200)
  @ApiOperation({ summary: 'Search all Restricao' })
  // @ApiOkResponse({ type: [any], description: 'The found Restricao' })
  async resultado(@Query() situation: string): Promise<any[]> {
    return await this.service.resultado({ situation });
  }

  // @Post('portarias')
  // @HttpCode(200)
  // @UseGuards(JwtAuthGuard, ACLGuard)
  // @ACLPolice({ roles: [], permissions: [] })
  // @ApiOperation({ summary: 'Found Restricao' })
  // @ApiOkResponse({
  //   type: CreateRestricaoDto,
  //   description: 'Found Restricao',
  // })
  // @ApiBadRequestResponse({ type: ErrorResponse, description: 'Bad Request' })
  // async findPortaria(@Body() data: SearchPortariaDto): Promise<any> {
  //   return await this.service.findPortaria(data);
  // }

  @Post()
  @HttpCode(201)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Create a new Restricao' })
  @ApiCreatedResponse({
    type: UpdateRestricaoDto,
    description: 'Created Restricao',
  })
  @ApiBadRequestResponse({ type: ErrorResponse, description: 'Bad Request' })
  async create(
    @Body() data: CreateRestricaoDto,
    @Request() request?: any,
  ): Promise<Restricao> {
    const response = await this.service.create(data);

    await activityLog({
      module: 'restricao',
      action: 'create',
      data: response,
      user: request?.user,
    });

    return response;
  }

  @Get(':id')
  @HttpCode(200)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Search a Restricao by id' })
  @ApiOkResponse({
    type: UpdateRestricaoDto,
    description: 'The found Restricao',
  })
  @ApiNotFoundResponse({ type: ErrorResponse, description: 'Not Found' })
  async findById(@Param('id') id: string): Promise<Restricao> {
    return await this.service.findById(id);
  }

  @Put(':id')
  @HttpCode(200)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Update a Restricao' })
  @ApiOkResponse({
    type: UpdateRestricaoDto,
    description: 'Updated Restricao',
  })
  @ApiNotFoundResponse({ type: ErrorResponse, description: 'Not Found' })
  async update(
    @Param('id') id: string,
    @Body() data: UpdateRestricaoDto,
    @Request() request?: any,
  ): Promise<Restricao> {
    const old = await this.service.findById(id);
    const response = await this.service.update(id, data);

    await activityLog({
      module: 'restricao',
      action: 'update',
      data: response,
      old,
      user: request?.user,
    });

    return response;
  }

  @Put(':id/restore')
  @HttpCode(200)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Delete a Restricao' })
  @ApiNoContentResponse({ description: 'Deleted Restricao' })
  @ApiNotFoundResponse({ type: ErrorResponse, description: 'Not Found' })
  async restore(
    @Param('id') id: string,
    @Request() request?: any,
  ): Promise<Restricao> {
    const data = await this.service.restore(id);

    await activityLog({
      module: 'restricao',
      action: 'restore',
      data,
      user: request?.user,
    });
    return data;
  }

  @Delete(':id')
  @HttpCode(200)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Delete a Restricao' })
  @ApiNoContentResponse({ description: 'Deleted Restricao' })
  @ApiNotFoundResponse({ type: ErrorResponse, description: 'Not Found' })
  async delete(
    @Param('id') id: string,
    @Request() request?: any,
  ): Promise<Restricao> {
    const data = await this.service.delete(id);

    await activityLog({
      module: 'restricao',
      action: 'delete',
      data,
      user: request?.user,
    });
    return data;
  }

  @Delete(':id/force')
  @HttpCode(200)
  @UseGuards(JwtAuthGuard, ACLGuard)
  @ACLPolice({ roles: [], permissions: [] })
  @ApiOperation({ summary: 'Delete definitive a Restricao' })
  @ApiNoContentResponse({ description: 'Deleted definitive Restricao' })
  @ApiNotFoundResponse({ type: ErrorResponse, description: 'Not Found' })
  async forceDelete(
    @Param('id') id: string,
    @Request() request?: any,
  ): Promise<Restricao> {
    const data = await this.service.forceDelete(id);

    await activityLog({
      module: 'restricao',
      action: 'forceDelete',
      data,
      user: request?.user,
    });
    return data;
  }
}
